+++ app.py
@@ -1,12 +1,4 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-
-@app.route('/')
-@app.route('/api/v1/hello-world-7')
-def index():
-    return 'Hello World-3'
+from pp_project import app
 
 
 if __name__ == "__main__":
