+++ pp_project/hello_world.py
@@ -1,0 +1,7 @@
+from pp_project import app
+
+
+@app.route('/')
+@app.route('/api/v1/hello-world-7')
+def index():
+    return 'Hello World-3'
